small fixes (#23)
* cast, fixes #8

* Arduino Server preferred API: ::available() -> ::accept()

* update to newer esp8266 ESP.getHeapStats api

+++ examples/Hooked/Hooked.ino
@@ -161,7 +161,7 @@ void help()
     Serial.printf("interactive: F/ormat D/ir C/reateFile\n");
 
     uint32_t freeHeap;
-    uint16_t maxBlock;
+    uint32_t maxBlock;
     uint8_t fragmentation;
     ESP.getHeapStats(&freeHeap, &maxBlock, &fragmentation);
     Serial.printf("Heap stats: free heap: %u - max block: %u - fragmentation: %u%%\n",
